requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.10.0=py37_0
aiohttp=3.6.3=py37h7b6447c_0
apptools=5.1.0=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py37_0
attrs=20.2.0=py_0
blas=1.0=mkl
blinker=1.4=py37_0
brotlipy=0.7.0=py37h7b6447c_1000
bzip2=1.0.8=h7b6447c_0
c-ares=1.16.1=h7b6447c_0
ca-certificates=2020.10.14=0
cachetools=4.1.1=py_0
cairo=1.16.0=h18b612c_1001
certifi=2020.6.20=py37_0
cffi=1.14.3=py37he30daa8_0
chardet=3.0.4=py37_1003
click=7.1.2=py_0
cloudpickle=1.6.0=py_0
cmake=3.14.0=h52cb24c_0
configobj=5.0.6=py37h06a4308_1
cryptography=3.1.1=py37h1ba5d50_0
csaps=0.11.0=py37_0
cudatoolkit=10.1.243=h6bb024c_0
cudatoolkit-dev=10.1.243=h516909a_3
cudnn=7.6.5=cuda10.1_0
cupti=10.1.168=0
cupy=8.3.0=py37hcaf9a05_0
curl=7.71.1=hbc83047_1
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.2.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pyhd3eb1b0_0
easydict=1.9=pypi_0
envisage=5.0.0=pyhd3eb1b0_0
expat=2.2.10=he6710b0_2
fastrlock=0.5=py37h3340039_1
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=py37_1
gast=0.3.3=py_0
giflib=5.2.1=h36c2ea0_2
glib=2.67.4=h36276a3_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=1.22.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.31.0=py37hf8bcb03_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=py37h7918eee_0
harfbuzz=2.4.0=h37c48d4_1
hdf4=4.2.13=h3ca952b_2
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=2.10=py_0
imageio=2.9.0=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
importlib_resources=5.1.1=py37h06a4308_1
intel-openmp=2020.2=254
jasper=1.900.1=h07fcdf6_1006
jpeg=9d=h36c2ea0_0
jsoncpp=1.8.4=hfd86e86_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.3.1=py37h2531618_0
krb5=1.18.2=h173b8e3_0
lame=3.100=h14c3975_1001
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=21_mkl
libcblas=3.8.0=21_mkl
libcurl=7.71.1=h20c2e04_1
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.16=h516909a_0
liblapack=3.8.0=21_mkl
liblapacke=3.8.0=21_mkl
libnetcdf=4.6.1=h11d0813_2
libogg=1.3.2=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=hd408876_0
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtheora=1.1.1=h5ab3b9f_1
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=h7b6447c_0
libwebp=1.0.2=h56121f0_5
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
llvmlite=0.36.0rc2=pypi_0
lz4-c=1.8.1.2=h14c3975_0
markdown=3.3.2=py37_0
matplotlib-base=3.3.4=py37h62a2d02_0
mayavi=4.7.1=py37h94891b3_2
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
multidict=4.7.6=py37h7b6447c_1
nccl=2.8.4.1=h8b44402_3
ncurses=6.2=he6710b0_1
nettle=3.6=he412f7d_0
networkx=2.5=py_0
ninja=1.10.2=py37hff7bd54_0
numba=0.53.0rc2=pypi_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
opencv=4.1.0=py37h5517eff_4
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1j=h27cfd23_0
opt_einsum=3.1.0=py_0
pcdet=0.3.0+0642cf0=dev_0
pcre=8.44=he6710b0_0
pillow=8.1.1=py37he98fc37_0
pip=21.0.1=py37h06a4308_0
pixman=0.38.0=h516909a_1003
protobuf=4.0.0rc2=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyface=7.3.0=py37h06a4308_1
pygments=2.8.0=pyhd3eb1b0_0
pyjwt=1.7.1=py37_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_1
python=3.7.10=hdb3f193_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.7=1_cp37m
pytorch=1.4.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.4.1=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
rhash=1.4.1=h3c74f83_1
rsa=4.6=py_0
scikit-image=0.17.2=py37hdf5156a_0
scipy=1.6.1=py37h91f5cce_0
setuptools=52.0.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
spconv=1.0=pypi_0
sqlite=3.33.0=h62c20be_0
tbb=2020.3=hfd86e86_0
tensorboard=2.2.1=pyh532a8cf_0
tensorboard-plugin-wit=1.6.0=py_0
tensorboardx=2.1=pypi_0
tensorflow=2.2.0=gpu_py37h1a511ff_0
tensorflow-base=2.2.0=gpu_py37h8a81be8_0
tensorflow-estimator=2.2.0=pyh208ff02_0
tensorflow-gpu=2.2.0=h0d30ee6_0
termcolor=1.1.0=py37_1
tifffile=2020.10.1=py37hdd07704_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
tornado=6.1=py37h27cfd23_0
tqdm=4.56.0=pyhd3eb1b0_0
traits=6.2.0=py37h27cfd23_0
traitsui=7.1.1=pyhd3eb1b0_0
urllib3=1.25.11=py_0
vtk=8.2.0=py37haa4764d_200
werkzeug=1.0.1=py_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=py37h7b6447c_1
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.2=h470a237_5
xorg-libx11=1.6.12=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.2=py37h7b6447c_0
yolov4=2.0.1=pypi_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
